Move KSP generators into endpoint and registry subpackages
