+++ helidon-extensions/src/main/kotlin/dev/sku20/helidon/ksp/endpoint/InitializersGenerator.kt
@@ -1,6 +1,9 @@
-package dev.sku20.helidon.ksp
+package dev.sku20.helidon.ksp.endpoint
 
 import com.google.devtools.ksp.symbol.KSClassDeclaration
+import dev.sku20.helidon.ksp.CustomWriter
+import dev.sku20.helidon.ksp.Utils
+import dev.sku20.helidon.ksp.registry.RegistryInitializerGenerator
 import java.io.InputStream
 import java.io.OutputStream
