+++ helidon-extensions/src/main/resources/META-INF/services/com.google.devtools.ksp.processing.SymbolProcessorProvider
@@ -1,1 +1,1 @@
-dev.sku20.helidon.ksp.EndpointSymbolProcessProvider
+dev.sku20.helidon.ksp.endpoint.EndpointSymbolProcessProvider
